refactor: PurchaseInformation에서 Money 분리
