+++ src/main/java/lotto/domain/Money.java
@@ -1,0 +1,23 @@
+package lotto.domain;
+
+public class Money {
+    private static final int LOTTO_PRICE = 1000;
+    private static final int MAX_LOTTO_PRICE = 1_000_000_000;
+
+    private final int money;
+
+    public Money(final int money) {
+        checkInvalidMoney(money);
+        this.money = money;
+    }
+
+    int calculateMaxNumberOfPurchase() {
+        return money / LOTTO_PRICE;
+    }
+
+    private void checkInvalidMoney(final int money) {
+        if (money < LOTTO_PRICE || money > MAX_LOTTO_PRICE) {
+            throw new IllegalArgumentException("로또 구입 가능 금액의 범위를 벗어났습니다.");
+        }
+    }
+}
